+++ .gitignore
@@ -1,6 +1,8 @@
 target/
 !.mvn/wrapper/maven-wrapper.jar
 
+application.properties
+
 ### STS ###
 .apt_generated
 .classpath
