+++ medly_db_roles.sql
@@ -8,9 +8,9 @@ INSERT INTO medly_db.roles (role) VALUES ('ROLE_DOCTOR');
 INSERT INTO medly_db.roles (role) VALUES ('ROLE_PATIENT');
 
 
-INSERT INTO users(account_verified, first_name, last_name, username, email, password, enabled, role_id)
-VALUES(1, 'joe', 'joe', 'test', 'test@test.com', '$2a$10$O5lrKNXfAdWWpf0P3Ij7Du03J7WuCRRtVDT7JGJwHTighGmweFhDO', 1, 3),
-      (1, 'tom', 'tom', 'root', 'tester@test.com', '$2a$10$mhGBce4Tu5TPLooNjVfOpObdX2tPblf/UKkLRniEtB4C.49EmrFb.', 1, 3);
+INSERT INTO users(first_name, last_name, username, email, password, enabled, role_id)
+VALUES('joe', 'joe', 'test', 'test@test.com', '$2a$10$O5lrKNXfAdWWpf0P3Ij7Du03J7WuCRRtVDT7JGJwHTighGmweFhDO', 1, 3),
+      ('tom', 'tom', 'root', 'tester@test.com', '$2a$10$mhGBce4Tu5TPLooNjVfOpObdX2tPblf/UKkLRniEtB4C.49EmrFb.', 1, 3);
 
 INSERT INTO medly_db.medications (brand_name, generic_name)
     VALUES ('Albuterol', 'ProAir® HFA'),
