+++ src/main/resources/application.properties
@@ -3,3 +3,7 @@ spring.datasource.url=jdbc:mysql://localhost/medly_db?createDatabaseIfNotExists=
 spring.datasource.username=root
 spring.datasource.password=codeup
 spring.jpa.hibernate.ddl-auto=update
+
+twillio-key=ACaff88bc37e7ed52a4980273b456d28ef
+twillio-token=893b20a6f30fe8de9b5e8031815ffdb2
+twillio-phone=+12103611492
