+++ src/main/resources/templates/fragments/registerForm.html
@@ -24,11 +24,11 @@
                 <!--<input th:field="*{dateOfBirth}" type="text" class="form-control" data-large-default="true"  data-fx-mobile="true" data-jump="4"  id="dateOfBirth" />-->
                 <!--<p th:if="${#fields.hasErrors('dateOfBirth')}" th:errors="*{dateOfBirth}"></p>-->
             <!--</div>-->
-            <!--<div class="form-group">-->
-                <!--<label for="phoneNumber">Phone Number:</label>-->
-                <!--<input th:field="*{phoneNumber}" type="text" class="form-control" id="phoneNumber"   />-->
-                <!--<p th:if="${#fields.hasErrors('phoneNumber')}" th:errors="*{phoneNumber}"></p>-->
-            <!--</div>-->
+            <div class="form-group">
+                <label for="phoneNumber">Phone Number:</label>
+                <input th:field="*{phoneNumber}" type="text" class="form-control" id="phoneNumber"   />
+                <p th:if="${#fields.hasErrors('phoneNumber')}" th:errors="*{phoneNumber}"></p>
+            </div>
             <!--<div class="form-group">-->
                 <!--<label for="streetAddress">Street Address:</label>-->
                 <!--<input th:field="*{streetAddress}" type="search" class="form-control" id="streetAddress" placeholder="Where do you live?"   />-->
